#9 - Extract Deserializer to its own file
