+++ src/test/kotlin/com/gmaur/investment/robotadvisor/infrastructure/DomainObjectMapperShould.kt
@@ -1,52 +1,16 @@
 package com.gmaur.investment.robotadvisor.infrastructure
 
-import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.core.JsonProcessingException
-import com.fasterxml.jackson.core.TreeNode
-import com.fasterxml.jackson.databind.DeserializationContext
-import com.fasterxml.jackson.databind.deser.std.StdDeserializer
-import com.fasterxml.jackson.databind.module.SimpleModule
-import com.fasterxml.jackson.databind.node.TextNode
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import com.fasterxml.jackson.module.kotlin.readValue
 import com.gmaur.investment.robotadvisor.domain.*
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.SoftAssertions
 import org.junit.Test
-import java.io.IOException
 import java.math.BigDecimal
 
 
 class DomainObjectMapperShould {
 
-    private val mapper = jacksonObjectMapper()
-
-    init {
-        val module = SimpleModule()
-        module.addDeserializer(XDTO::class.java, XDeserializer())
-        mapper.registerModule(module)
-    }
-
-    inner class XDeserializer @JvmOverloads constructor(vc: Class<*>? = null) : StdDeserializer<XDTO>(vc) {
-
-        @Throws(IOException::class, JsonProcessingException::class)
-        override fun deserialize(jp: JsonParser, ctxt: DeserializationContext): XDTO {
-            val node = jp.codec.readTree<TreeNode>(jp)
-            val result = when (string(node, "type")) {
-                "fund" -> {
-                    XDTO(AssetDTO(isin = string(node, "isin"), transferrable = false), amount = AmountDTO.EUR(string(node, "price")))
-                }
-                "cash" -> {
-                    XDTO(AssetDTO(isin = "", transferrable = true), amount = AmountDTO.EUR(string(node, "value")))
-                }
-                else -> {
-                    throw IllegalArgumentException("type not recognized in: " + node.toString())
-                }
-            }
-            return result
-        }
-
-        private fun string(node: TreeNode, key: String) = (node.get(key) as TextNode).asText()
-    }
+    private val mapper = JSONMapper.aNew()
 
     @Test
     fun `convert from domain to DTO`() {
@@ -107,6 +71,7 @@ class DomainObjectMapperShould {
         softly.assertAll()
     }
 
+
     @Test
     fun `parse from the other domain`() {
         val otherDomainJson = """
@@ -133,7 +98,7 @@ class DomainObjectMapperShould {
   } ]
 }
 """
-        val dto = mapper.readValue<PortfolioDTO>(otherDomainJson, PortfolioDTO::class.java)
+        val dto = mapper.readValue<PortfolioDTO>(otherDomainJson)
 
         var domain = DomainObjectMapper().toDomain(dto)
